# backend/src/main/java/br/com/tradewind/dto/OrderItemRequest.java
package br.com.tradewind.dto;
import jakarta.validation.constraints.Min;
import jakarta.validation.constraints.NotNull;
import java.util.UUID;
public record OrderItemRequest(@NotNull UUID productId, @Min(1) int quantity) { }