# docs/postman/openapi.json
{"openapi":"3.1.0","info":{"title":"API Tradewind","description":"API de referencia do template tradewind-fullstack.\n\nRoda contra o Postgres do Docker Compose (perfil `local`) ou\ncontra o Supabase (perfil `supabase`) usando o mesmo schema\ngerenciado por Flyway.\n","contact":{"name":"Equipe Tradewind","email":"dev@tradewind.local"},"license":{"name":"MIT"},"version":"1.0.0"},"servers":[{"url":"http://localhost:8080","description":"Local / Docker"},{"url":"https://api.tradewind.exemplo.com","description":"Producao"}],"tags":[{"name":"Estoque","description":"Saldo por produto e localizacao"},{"name":"Relatorios","description":"Indicadores do dashboard"},{"name":"Produtos","description":"CRUD do catalogo de produtos"},{"name":"Pedidos","description":"Checkout acadêmico com estoque e pagamento simulado"},{"name":"Clientes","description":"Cadastro minimo, exportacao e anonimizacao"}],"paths":{"/api/v1/products/{id}":{"get":{"tags":["Produtos"],"summary":"Busca produto por id","operationId":"getById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"Produto encontrado","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductResponse"}}}},"404":{"description":"Produto inexistente","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductResponse"}}}}}},"put":{"tags":["Produtos"],"summary":"Atualiza produto","operationId":"update","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductRequest"}}},"required":true},"responses":{"200":{"description":"Produto atualizado","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductResponse"}}}},"404":{"description":"Produto inexistente","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductResponse"}}}},"409":{"description":"Sku ja usado por outro produto","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductResponse"}}}}}},"delete":{"tags":["Produtos"],"summary":"Remove produto","operationId":"delete","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"204":{"description":"Produto removido"},"404":{"description":"Produto inexistente"}}}},"/api/v1/products":{"get":{"tags":["Produtos"],"summary":"Lista produtos","description":"Filtros sao combinados; todos sao opcionais.","operationId":"list","parameters":[{"name":"term","in":"query","description":"Busca por nome ou sku","required":false,"schema":{"type":"string"}},{"name":"active","in":"query","description":"Filtra por disponibilidade","required":false,"schema":{"type":"boolean"}},{"name":"minPrice","in":"query","description":"Preco minimo","required":false,"schema":{"type":"number"}},{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":20}}],"responses":{"200":{"description":"Pagina de produtos","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PageResponse"}}}}}},"post":{"tags":["Produtos"],"summary":"Cria produto","operationId":"create","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductRequest"}}},"required":true},"responses":{"201":{"description":"Produto criado","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductResponse"}}}},"400":{"description":"Payload invalido","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductResponse"}}}},"409":{"description":"Sku ja cadastrado","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductResponse"}}}}}}},"/api/v1/orders":{"post":{"tags":["Pedidos"],"summary":"Cria pedido e reserva estoque","operationId":"create_1","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateOrderRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrderResponse"}}}}}}},"/api/v1/orders/{id}/payments":{"post":{"tags":["Pedidos"],"summary":"Simula resultado de pagamento, sem dados financeiros","operationId":"pay","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SimulatePaymentRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PaymentResponse"}}}}}}},"/api/v1/inventory/{id}/adjust":{"post":{"tags":["Estoque"],"summary":"Ajusta saldo manualmente","description":"Quantidade positiva repõe, negativa consome. Todo ajuste vira movimentacao com motivo.","operationId":"adjust","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/StockAdjustmentRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/InventoryView"}}}}}}},"/api/v1/customers":{"post":{"tags":["Clientes"],"summary":"Cadastra cliente","description":"Coleta minima: e-mail, nome, locale e pais. Consentimento de marketing e falso por tradewind.","operationId":"create_2","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CustomerRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CustomerView"}}}}}}},"/api/v1/reports/declined-payments":{"get":{"tags":["Relatorios"],"summary":"Pagamentos recusados","description":"Lista os eventos de recusa, que e o sinal monitorado pelo modulo antifraude simulado.","operationId":"declinedPayments","parameters":[{"name":"limit","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":50}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/DeclinedPayment"}}}}}}}},"/api/v1/reports/dashboard":{"get":{"tags":["Relatorios"],"summary":"Painel consolidado","description":"Vendas, receita por moeda, produtos mais vendidos, conversao e estoque em alerta.","operationId":"dashboard","parameters":[{"name":"days","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":30}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Dashboard"}}}}}}},"/api/v1/products/by-sku/{sku}":{"get":{"tags":["Produtos"],"summary":"Busca produto por sku","operationId":"getBySku","parameters":[{"name":"sku","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductResponse"}}}}}}},"/api/v1/orders/{id}":{"get":{"tags":["Pedidos"],"summary":"Consulta pedido","operationId":"get","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrderResponse"}}}}}}},"/api/v1/inventory":{"get":{"tags":["Estoque"],"summary":"Lista saldos","description":"`onlyLowStock=true` devolve so o que esta no ou abaixo do ponto de reposicao.","operationId":"list_1","parameters":[{"name":"onlyLowStock","in":"query","required":false,"schema":{"type":"boolean","default":false}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/InventoryView"}}}}}}}},"/api/v1/inventory/forecast":{"get":{"tags":["Estoque"],"summary":"Previsao de demanda","description":"Media movel das saidas confirmadas por janela. Sugere reposicao e dias de cobertura.","operationId":"forecast","parameters":[{"name":"windowDays","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":7}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ForecastRow"}}}}}}}},"/api/v1/customers/{id}":{"get":{"tags":["Clientes"],"summary":"Consulta cadastro","operationId":"get_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CustomerView"}}}}}},"delete":{"tags":["Clientes"],"summary":"Anonimiza o titular","description":"Preserva os pedidos, que sao registro contabel; apaga o que identifica a pessoa.","operationId":"anonymize","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK"}}}},"/api/v1/customers/{id}/export":{"get":{"tags":["Clientes"],"summary":"Exporta os dados do titular","description":"Portabilidade exigida por LGPD art. 18 V e GDPR art. 20","operationId":"export","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CustomerExport"}}}}}}}},"components":{"schemas":{"ProductRequest":{"type":"object","description":"Payload de criacao/atualizacao de produto","properties":{"sku":{"type":"string","description":"Codigo de estoque (unico)","example":"SKU-001","maxLength":64,"minLength":0},"name":{"type":"string","description":"Nome comercial","example":"Cafe Torrado 500g","maxLength":255,"minLength":0},"description":{"type":"string","description":"Descricao livre","example":"Torra media, notas de chocolate","maxLength":1000,"minLength":0},"price":{"type":"number","description":"Preco unitario","example":39.9,"minimum":0.00},"active":{"type":"boolean","default":"true","description":"Disponivel na loja","example":true}},"required":["name","price","sku"]},"ProductResponse":{"type":"object","description":"Produto retornado pela API","properties":{"id":{"type":"string","format":"uuid","description":"Identificador do produto","example":"9f1c2b7e-6d2a-4c0e-9a3b-8f2d5c1e7a40"},"sku":{"type":"string","example":"SKU-001"},"name":{"type":"string","example":"Cafe Torrado 500g"},"description":{"type":"string","example":"Torra media, notas de chocolate"},"price":{"type":"number","example":39.9},"active":{"type":"boolean","example":true},"createdAt":{"type":"string","format":"date-time","example":"2026-09-25T12:00:00-03:00"}}},"CreateOrderRequest":{"type":"object","properties":{"customerEmail":{"type":"string","maxLength":320,"minLength":0},"customerName":{"type":"string","maxLength":160,"minLength":0},"locale":{"type":"string","pattern":"[a-z]{2}(-[A-Z]{2})?"},"countryCode":{"type":"string","pattern":"[A-Z]{2}"},"currency":{"type":"string","pattern":"[A-Z]{3}"},"items":{"type":"array","items":{"$ref":"#/components/schemas/OrderItemRequest"},"minItems":1}}},"OrderItemRequest":{"type":"object","properties":{"productId":{"type":"string","format":"uuid"},"quantity":{"type":"integer","format":"int32","minimum":1}},"required":["productId"]},"Item":{"type":"object","properties":{"productId":{"type":"string","format":"uuid"},"sku":{"type":"string"},"name":{"type":"string"},"quantity":{"type":"integer","format":"int32"},"unitPrice":{"type":"number"},"lineTotal":{"type":"number"}}},"OrderResponse":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"orderNumber":{"type":"string"},"status":{"type":"string","enum":["AWAITING_PAYMENT","CONFIRMED","PAYMENT_PENDING","PAYMENT_DECLINED","CANCELLED"]},"currency":{"type":"string"},"subtotal":{"type":"number"},"taxTotal":{"type":"number"},"grandTotal":{"type":"number"},"countryCode":{"type":"string"},"locale":{"type":"string"},"createdAt":{"type":"string","format":"date-time"},"items":{"type":"array","items":{"$ref":"#/components/schemas/Item"}}}},"SimulatePaymentRequest":{"type":"object","properties":{"method":{"type":"string","enum":["PIX","CARD","BANK_TRANSFER","DIGITAL_WALLET"]},"outcome":{"type":"string","enum":["APPROVED","DECLINED","PENDING"]}},"required":["method","outcome"]},"PaymentResponse":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"orderId":{"type":"string","format":"uuid"},"method":{"type":"string","enum":["PIX","CARD","BANK_TRANSFER","DIGITAL_WALLET"]},"status":{"type":"string","enum":["APPROVED","DECLINED","PENDING"]},"amount":{"type":"number"},"currency":{"type":"string"},"simulationReference":{"type":"string"},"processedAt":{"type":"string","format":"date-time"}}},"StockAdjustmentRequest":{"type":"object","properties":{"quantity":{"type":"integer","format":"int32","maximum":1000000,"minimum":-1000000},"reason":{"type":"string","maxLength":255,"minLength":0}},"required":["quantity"]},"InventoryView":{"type":"object","description":"Saldo de um produto em uma localizacao","properties":{"id":{"type":"string","format":"uuid"},"productId":{"type":"string","format":"uuid"},"sku":{"type":"string"},"productName":{"type":"string"},"locationId":{"type":"string","format":"uuid"},"locationCode":{"type":"string"},"availableQuantity":{"type":"integer","format":"int32"},"reservedQuantity":{"type":"integer","format":"int32"},"reorderPoint":{"type":"integer","format":"int32"},"lowStock":{"type":"boolean"}}},"CustomerRequest":{"type":"object","description":"Cadastro de cliente. Coleta minima: sem documento, telefone ou endereco.","properties":{"email":{"type":"string","maxLength":320,"minLength":0},"displayName":{"type":"string","maxLength":160,"minLength":0},"locale":{"type":"string","pattern":"[a-z]{2}(-[A-Z]{2})?"},"countryCode":{"type":"string","pattern":"[A-Z]{2}"}}},"CustomerView":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"email":{"type":"string"},"displayName":{"type":"string"},"locale":{"type":"string"},"countryCode":{"type":"string"},"marketingConsent":{"type":"boolean"}}},"DeclinedPayment":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"orderNumber":{"type":"string"},"method":{"type":"string","enum":["PIX","CARD","BANK_TRANSFER","DIGITAL_WALLET"]},"amount":{"type":"number"},"currency":{"type":"string"},"processedAt":{"type":"string","format":"date-time"}}},"CountrySales":{"type":"object","properties":{"countryCode":{"type":"string"},"orders":{"type":"integer","format":"int64"},"currency":{"type":"string"},"revenue":{"type":"number"}}},"Dashboard":{"type":"object","properties":{"windowDays":{"type":"integer","format":"int32"},"ordersCreated":{"type":"integer","format":"int32"},"ordersConfirmed":{"type":"integer","format":"int32"},"revenueByCurrency":{"type":"object","additionalProperties":{"type":"number"}},"salesByCountry":{"type":"array","items":{"$ref":"#/components/schemas/CountrySales"}},"topProducts":{"type":"array","items":{"$ref":"#/components/schemas/ProductSales"}},"conversionRatePercent":{"type":"number","format":"double"},"pendingPayments":{"type":"integer","format":"int64"},"lowStockRows":{"type":"integer","format":"int64"},"declinedOrders":{"type":"integer","format":"int64"}}},"ProductSales":{"type":"object","properties":{"sku":{"type":"string"},"productName":{"type":"string"},"currency":{"type":"string"},"units":{"type":"integer","format":"int32"},"revenue":{"type":"number"},"unitPrice":{"type":"number"}}},"PageResponse":{"type":"object","description":"Pagina de resultados","properties":{"content":{"type":"array","description":"Itens da pagina atual","items":{"type":"object"}},"page":{"type":"integer","format":"int32","description":"Numero da pagina (base 0)","example":0},"size":{"type":"integer","format":"int32","description":"Tamanho da pagina","example":20},"totalElements":{"type":"integer","format":"int64","description":"Total de itens em todos os filtros","example":5},"totalPages":{"type":"integer","format":"int32","description":"Total de paginas","example":1},"first":{"type":"boolean","description":"Primeira pagina?","example":true},"last":{"type":"boolean","description":"Ultima pagina?","example":true}}},"ForecastRow":{"type":"object","properties":{"sku":{"type":"string"},"productName":{"type":"string"},"locationCode":{"type":"string"},"unitsSoldInWindow":{"type":"integer","format":"int32"},"dailyRate":{"type":"number","format":"double"},"availableQuantity":{"type":"integer","format":"int32"},"reorderPoint":{"type":"integer","format":"int32"},"suggestedReorder":{"type":"integer","format":"int32"},"daysOfCover":{"type":"integer","format":"int32"}}},"CustomerExport":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"email":{"type":"string"},"displayName":{"type":"string"},"locale":{"type":"string"},"countryCode":{"type":"string"},"orders":{"type":"array","items":{"$ref":"#/components/schemas/OrderSummary"}}}},"OrderSummary":{"type":"object","properties":{"orderNumber":{"type":"string"},"status":{"type":"string","enum":["AWAITING_PAYMENT","CONFIRMED","PAYMENT_PENDING","PAYMENT_DECLINED","CANCELLED"]},"currency":{"type":"string"},"grandTotal":{"type":"number"},"createdAt":{"type":"string","format":"date-time"}}}}}}